split MultiStack creation from HostNavigator creation
